Update .ranks_around_position to respect scope_by (#12)
In the initial version, using a scope by would not effect the columns
returned when calculating the before and after ranks for the position.
This would cause problems when you expected rankings to be scoped to a
column value, since `.ranks_around_position` would return all records,
and could possibly cause an error when moving to a rank that didn't make
sense in the concept of the scope, or ever worse, succeed in re-ranking
the element, but in the wrong place.

`.ranks_around_position` now takes an optional `scope_value` keyword
(defaults to nil) to be used in the where clause of the query to find
the previous and next ranking values when moving a record to an absolute
position.

This changes how the public methods `.ranks_around_position` and
`#move_to` work, but in a backward compatable way. Thus the version bump
to 0.2.0.